+++ ps2_cam_acolyte/games/shared_ui.h
@@ -26,6 +26,16 @@ public:
 		ImGui::TextColored(toggle.is_on() ? ui_colors::on_obvious : ui_colors::off_obvious, toggle.is_on() ? "ON" : "OFF");
 	}
 
+	static void slider(preferences& prefs, const char* label, const char* pref, float* value, float min_value = 0.0f, float max_value = 1.0f)
+	{
+		float initial_value = *value;
+		ImGui::SliderFloat(label, value, min_value, max_value, "%.2f");
+		if (*value != initial_value)
+		{
+			prefs.write_float(pref, *value);
+		}
+	}
+
 	static void button(const controller& c, button_type button)
 	{
 		ImGui::Text("[%s]", c.get_button_display_name(button));