+++ ps2_cam_acolyte/preferences.h
@@ -15,4 +15,8 @@ public:
 	void write(const char* key, const char* value);
 	float read_float(const char* key, float default_value) const;
 	void write_float(const char* key, float value);
+	void write_on_change_float(const char* key, float* value, float target_value);
+	bool read_bool(const char* key, bool default_value) const;
+	void write_bool(const char* key, bool value);
+	void write_on_change_bool(const char* key, bool* value, bool target_value);
 };